feat(import): resolve E2E tests, refactor ImportNotificationBanner sub-components for ESLint limits, and sync eslint whitelist
