+++ eslint.config.mjs
@@ -126,8 +126,6 @@ export default ts.config(
   {
     files: [
       "src/placeholder-non-existent-max-lines.ts",
-      "src/components/organisms/HandBar.tsx",
-      "src/components/organisms/EasyModeSection.tsx",
       "src/components/organisms/Workbench.tsx",
     ],
     rules: {
@@ -139,7 +137,6 @@ export default ts.config(
     files: [
       "src/placeholder-non-existent-complexity.ts",
       "src/components/organisms/CloudSyncSection.tsx",
-      "src/components/templates/SidePanelLayout.tsx",
       "src/lib/prompt-utils.ts",
     ],
     rules: {
