+++ playwright.config.ts
@@ -7,18 +7,18 @@ import { defineConfig, devices } from "@playwright/test"
 export default defineConfig({
   testDir: "./tests/e2e",
   /* Maximum time one test can run for. */
-  timeout: 30 * 1000,
+  timeout: 60 * 1000,
   expect: {
-    timeout: 5000
+    timeout: 10000
   },
-  /* Run tests in files in parallel */
-  fullyParallel: true,
+  /* Run tests sequentially to reduce machine load and flakes */
+  fullyParallel: false,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
   /* Opt out of parallel tests on CI. */
-  workers: process.env.CI ? 2 : undefined,
+  workers: 1,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: "list",
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
@@ -46,8 +46,7 @@ export default defineConfig({
 
   /* Run local dev server before starting the tests */
   webServer: {
-    command:
-      `"${process.execPath}" ./node_modules/vite/bin/vite.js --config tests/sandbox/vite.config.ts --port 5173 --strictPort`,
+    command: `"${process.execPath}" ./node_modules/vite/bin/vite.js --config tests/sandbox/vite.config.ts --port 5173 --strictPort`,
     port: 5173,
     reuseExistingServer: !process.env.CI,
     timeout: 15 * 1000
