+++ src/locales/ja/translation.json
@@ -265,7 +265,12 @@
     "importing": "カードをインポート中...",
     "importSuccess": "カード「{name}」を正常にインポートしました！",
     "associated": "カード「{name}」に関連付けられました！",
-    "historyAdded": "新しい履歴アイテムが追加されました！"
+    "historyAdded": "新しい履歴アイテムが追加されました！",
+    "errorNoMetadataReason": "アップロードされた画像からスタイル情報（メタデータやQRコード）が見つかりませんでした。SNSへのアップロード時などに画像が圧縮され、データが削られた可能性があります。",
+    "errorNoMetadataActionTitle": "解決のための3つのステップ：",
+    "errorNoMetadataStep1": "元画像をインポートする：SNS等を経由していない、カメラで撮影したか直接保存したオリジナルのPNG画像ファイルをドラッグ＆ドロップしてください。",
+    "errorNoMetadataStep2": "QRコード部分をトリミングしてインポートする：画像内のQRコード部分だけを切り抜いた（トリミングした）画像をドロップしてみてください。",
+    "errorNoMetadataStep3": "プロンプトを手動で登録する：カード画像を見ながら、「ワークベンチ」タブから手動でプロンプトを入力して新規作成してください。"
   },
   "share": {
     "shareTitle": "スタイルカードの共有",
